Add host callbacks to component model
Target should be to define wit signatures with de C api, and pass that
along with the callback itself.

For now, this is a simplified version tailored to secretarium use,
where we define a much reduced set of allowed signatures with a
wamr-like string : only s32, strings and list<u8> for parameters and
return values.

This makes it reasonably easy to share the same callbacks whether in a
component model context of for plain wasm modules.